Add files via upload
The "Smart Helmet for Accident Detection" is an IoT-based safety device designed to detect accidents involving bike riders and notify emergency services with the rider’s location. The system utilizes an MPU6050 accelerometer to detect sudden impacts or abnormal tilting and a vibration sensor to confirm the shock. A GPS module provides real-time location data, and a GSM module sends an emergency SMS with the location to a predefined contact. A buzzer and LED offer alerts, and a cancel button allows users to stop false alarms. This system enhances road safety by enabling quick response to accidents.